ai/pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
	"flask>=3.1.3",
	"google-genai>=1.66.0",
	"python-dotenv>=1.2.2",
]
